Update MultiClient.Net to version 4.

+++ RogueCastle/SaveGameManager.cs
@@ -1,13 +1,11 @@
+// Rogue Legacy Randomizer - SaveGameManager.cs
+// Last Modified 2022-10-24
 // 
-//  Rogue Legacy Randomizer - SaveGameManager.cs
-//  Last Modified 2022-01-26
+// This project is based on the modified disassembly of Rogue Legacy's engine, with permission to do so by its
+// original creators. Therefore, the former creators' copyright notice applies to the original disassembly.
 // 
-//  This project is based on the modified disassembly of Rogue Legacy's engine, with permission to do so by its
-//  original creators. Therefore, the former creators' copyright notice applies to the original disassembly.
-// 
-//  Original Source - © 2011-2015, Cellar Door Games Inc.
-//  Rogue Legacy™ is a trademark or registered trademark of Cellar Door Games Inc. All Rights Reserved.
-// 
+// Original Source © 2011-2015, Cellar Door Games Inc.
+// Rogue Legacy™ is a trademark or registered trademark of Cellar Door Games Inc. All Rights Reserved.
 
 using System;
 using System.Collections.Generic;
@@ -1734,8 +1732,8 @@ namespace RogueCastle
                     {
                         var item = new NetworkItem
                         {
-                            Item = binaryReader.ReadInt32(),
-                            Location = binaryReader.ReadInt32(),
+                            Item = binaryReader.ReadInt64(),
+                            Location = binaryReader.ReadInt64(),
                             Player = binaryReader.ReadInt32()
                         };
